requirements.txt
# requirements.txt
sounddevice>=0.4.6
numpy>=1.24.0
soundfile>=0.12.0
groq>=0.3.0
webrtcvad>=2.0.10
python-dotenv>=1.0.0
torch==2.0.1+cu118
torchvision==0.15.2+cu118
torchaudio==2.0.2+cu118 --index-url https://download.pytorch.org/whl/cu118
accelerate bitsandbytes
nest_asyncio